requirements.txt
python-telegram-bot==20.6
APScheduler==3.11.0
SQLAlchemy==2.0.40
dateparser==1.2.0
openai==0.28
psycopg2-binary